+++ engine/engine/server/engine_server_service.go
@@ -38,7 +38,7 @@ func (service *EngineServerService) CreateEnclave(ctx context.Context, args *kur
 		return nil, stacktrace.Propagate(err, "An error occurred parsing the log level string '%v':", args.ApiContainerLogLevel)
 	}
 
-	enclave, err := service.enclaveManager.CreateEnclave(
+	enclaveInfo, err := service.enclaveManager.CreateEnclave(
 		ctx,
 		args.ApiContainerImage,
 		apiContainerLogLevel,
@@ -51,33 +51,18 @@ func (service *EngineServerService) CreateEnclave(ctx context.Context, args *kur
 	}
 
 	response := &kurtosis_engine_rpc_api_bindings.CreateEnclaveResponse{
-		NetworkId:                   enclave.GetNetworkId(),
-		NetworkCidr:                 enclave.GetNetworkIpAndMask().String(),
-		ApiContainerId:              enclave.GetApiContainerId(),
-		ApiContainerIpInsideNetwork: enclave.GetApiContainerIpAddr().String(),
-		ApiContainerHostIp:          enclave.GetApiContainerHostPortBinding().HostIP,
-		ApiContainerHostPort:        enclave.GetApiContainerHostPortBinding().HostPort,
+		EnclaveInfo: enclaveInfo,
 	}
 
 	return response, nil
 }
 
-func (service *EngineServerService) GetEnclave(ctx context.Context, args *kurtosis_engine_rpc_api_bindings.GetEnclaveArgs) (*kurtosis_engine_rpc_api_bindings.GetEnclaveResponse, error) {
-
-	enclave, err := service.enclaveManager.GetEnclave(ctx, args.EnclaveId)
+func (service *EngineServerService) GetEnclaves(ctx context.Context, _ *emptypb.Empty) (*kurtosis_engine_rpc_api_bindings.GetEnclavesResponse, error) {
+	infoForEnclaves, err := service.enclaveManager.GetEnclaves(ctx)
 	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting enclave with ID '%v'", args.EnclaveId)
+		return nil, stacktrace.Propagate(err, "An error occurred getting info for enclaves")
 	}
-
-	response := &kurtosis_engine_rpc_api_bindings.GetEnclaveResponse{
-		NetworkId:                   enclave.GetNetworkId(),
-		NetworkCidr:                 enclave.GetNetworkIpAndMask().String(),
-		ApiContainerId:              enclave.GetApiContainerId(),
-		ApiContainerIpInsideNetwork: enclave.GetApiContainerIpAddr().String(),
-		ApiContainerHostIp:          enclave.GetApiContainerHostPortBinding().HostIP,
-		ApiContainerHostPort:        enclave.GetApiContainerHostPortBinding().HostPort,
-	}
-
+	response := &kurtosis_engine_rpc_api_bindings.GetEnclavesResponse{EnclaveInfo: infoForEnclaves}
 	return response, nil
 }